Add supplemental schema to cps.py
Some attributes supported by CPS are part of the supplemental schema. These
attributes were not saved in the object created from the CPS file. This commit
add support for these attributes that are already documented as part of CPS.
If any or all of these attributes are not available in the CPS file, they
are simply ignored.