Convert compilation flags to CPS attributes
`-l` and `-L` flags are recognized and values given through these
flags are converted to known CPS attributes, `Requires` and `Location`.

The value of the `Location` attribute is computed using the information
contained in `-L` (path in which the library is searched for) and
the library name. The library name can be given with a new command line
option `--library-name`. If it is not specified, it will be assumed that
the name of the library is the base name of the input pkgconfig file.

The flags `-l` and `-L` are removed from the list of flags saved under
`Link-Flags` except if some libraries are marked as black-listed.

The library searched for in the path given by the flags `-L` is assumed
to have one of the following file extension: `.a`, `.so`, `.dylib`. If
it is a shared object, it can also contain a version number.